+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/template/ToggleControlsTemplateContextContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.admin.web.internal.theme.contributor;
+package com.liferay.layout.admin.web.internal.template;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.Layout;
