+++ modules/apps/product-navigation/product-navigation-control-menu-theme-contributor/src/main/java/com/liferay/product/navigation/control/menu/theme/contributor/internal/template/ProductNavigationControlMenuTemplateContextContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.product.navigation.control.menu.theme.contributor.internal;
+package com.liferay.product.navigation.control.menu.theme.contributor.internal.template;
 
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.template.TemplateContextContributor;
