+++ modules/apps/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/internal/template/UsersTemplateContextContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.users.admin.web.internal.theme.contributor;
+package com.liferay.users.admin.web.internal.template;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
